+++ packages/cli/src/cliBase.ts
@@ -3,10 +3,11 @@ import minimist from "minimist";
 import {
   ask,
   baseCommands,
+  createEnv,
+  DEFAULT_ENVS,
   displaySpinnerAsync,
   handle,
   loadDefaultEnv,
-  migrateLegacyEnv,
   parseInput,
   State,
   title,
@@ -17,6 +18,7 @@ const inquirerCommandPrompt = require("inquirer-command-prompt");
 const InterruptedPrompt = require("inquirer-interrupted-prompt");
 
 import AutocompletePrompt from "inquirer-autocomplete-prompt";
+import { GQL_URLS } from "@andromedaprotocol/andromeda.js";
 class CustomAutocompletePrompt extends AutocompletePrompt<any> {
   onSubmit(line: string) {
     let selected_value = this.currentChoices.getChoice(this.selected)?.value;
@@ -34,7 +36,12 @@ inquirer.registerPrompt("command", inquirerCommandPrompt);
 
 async function onStartup() {
   try {
-    migrateLegacyEnv();
+    const environments = ['TESTNET', 'MAINNET', 'DEVNET'] as const;
+    environments.forEach(env => {
+      createEnv(DEFAULT_ENVS[env], {
+        'gql': GQL_URLS[env],
+      }, true);
+    })
     await displaySpinnerAsync("Loading env..", loadDefaultEnv);
   } catch (error) {
     console.error(error);
@@ -46,7 +53,7 @@ export async function start() {
   const inputs = process.argv.slice(2);
   if (inputs.length === 0) {
     await title();
-    await State.connectClient()
+    await State.connectClient().catch(_ => { })
     while (true) {
       let input = await ask();
       const { _: cmd, ...flags } = minimist(parseInput(input.command));