+++ packages/cli/src/handlers/wasm.ts
@@ -1,6 +1,5 @@
-import { Msg } from "@andromedaprotocol/andromeda.js";
+import { convertMicroToMacro, getUnitsConfigFromDenom, Msg, formatAmountToInternational } from "@andromedaprotocol/andromeda.js";
 import { Coin, parseCoins } from "@cosmjs/proto-signing";
-import { StdFee } from "@cosmjs/stargate";
 import { promptWithExit } from "../cmd";
 import fs from "fs";
 import path from "path";
@@ -16,7 +15,7 @@ import State from "../state";
 import { Commands, Flags } from "../types";
 import { parseJSONInput, validateAddressInput } from "./utils";
 import Table from "cli-table";
-
+import { StdFee } from "@cosmjs/amino";
 
 export const commands: Commands = {
   'info': {
@@ -66,13 +65,25 @@ export const commands: Commands = {
         requestMessage: "Input Contract Address:",
         validate: validateAddressInput,
       },
-      {
-        requestMessage: "Input key:",
-        validate: (input: string) => {
-          return input.length > 0;
-        },
-      },
     ],
+    flags: {
+      limit: {
+        description: "Limit number of results",
+        usage: "--limit 10",
+      },
+      offset: {
+        description: "Offset number of results (This might not be supported by all chains)",
+        usage: "--offset 0",
+      },
+      ['next-key']: {
+        description: "Next key to paginate from",
+        usage: "--next-key <key>",
+      },
+      ['next-key-bytes']: {
+        description: "Next key to paginate from (bytes)",
+        usage: "--next-key-bytes <key>",
+      }
+    }
   },
   execute: {
     handler: executeHandler,
@@ -251,18 +262,41 @@ async function queryHandler(input: string[]) {
  * Queries a contract given a query message and address
  * @param input
  */
-async function queryRawHandler(input: string[]) {
-  const [contractAddr, key] = input;
-  const resp = await displaySpinnerAsync(
-    "Querying contrac key...",
-    async () => await State.client.queryContractRaw<any>(contractAddr, key)
-  );
-  if (resp === null) {
-    console.log(pc.red("No data found at key"));
-    return;
+async function queryRawHandler(input: string[], flags: Flags) {
+  const [contractAddr] = input;
+  const limit = BigInt(flags.limit ?? '10');
+  const offset = BigInt(flags.offset ?? '0');
+  const nextKey = flags['next-key'] as string | undefined;
+  const nextKeyBytes = flags['next-key-bytes'] as string | undefined;
+
+  if (nextKeyBytes && nextKey) {
+    throw new Error("Cannot provide both next-key and next-key-bytes");
   }
-  console.log(pc.green("Response: "));
-  console.log(JSON.stringify(resp, null, 2));
+
+  const key = nextKeyBytes ? Uint8Array.from(Buffer.from(nextKeyBytes, 'hex')) : nextKey ? Uint8Array.from(Buffer.from(nextKey, 'utf8')) : undefined;
+
+  const states = await displaySpinnerAsync(
+    "Querying contrac key...",
+    async () => await State.client.queryContractRawAll(contractAddr, { limit, offset, key })
+  );
+  console.log();
+
+
+  for (const state of states.states) {
+    console.log(pc.blue(state.key));
+    console.log(state.value);
+    console.log();
+  }
+  console.log();
+  console.log(pc.gray(`Limit - ${limit}, Offset - ${offset}, key - ${key ? Buffer.from(key).toString('utf8') : 'None'}`))
+  if (states.pagination) {
+    console.log(pc.gray(`Next Key - ${Buffer.from(states.pagination.nextKey).toString('utf8')}`));
+    console.log(pc.gray(`Next Key Bytes - ${Buffer.from(states.pagination.nextKey).toString('hex')}`));
+    console.log(pc.gray(`Total - ${states.pagination.total.toString()}`));
+  }
+  console.log()
+
+
 }
 
 /**
@@ -580,7 +614,9 @@ function logFeeEstimation(fee: StdFee) {
   console.log("Fee estimates:");
   for (let i = 0; i < fee.amount.length; i++) {
     const feeCoin = fee.amount[i];
-    console.log(`   ${pc.green(`${feeCoin.amount}${feeCoin.denom}`)}`);
+    const denomUnits = getUnitsConfigFromDenom(feeCoin.denom);
+    const macroAmount = convertMicroToMacro(feeCoin.amount, denomUnits.units);
+    console.log(`   ${pc.green(`${formatAmountToInternational(macroAmount)} ${denomUnits.macroDenom}`)}`);
   }
   console.log();
 }