+++ packages/adocodegen/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @andromedaprotocol/adocodegen
 
+## 1.0.1-beta.0
+
+### Patch Changes
+
+- Fix: Ether wallet now apply Bech32 prefix properly
+
 ## 1.0.0
 
 ### Major Changes